:truck: Move coffee menu related files into coffee package
